Fixes Reactor Crash
Fixes the possibility of the energy network crashing and the energy regulator disappearing.  This could have been caused by a Null pointer Exception on previous line 291 (now 294), where the BlockStorage checks for the stored "reactor-mode", when if the mode was never changed it would not have been set at all.