Only create a hopper "target" as long as the block is actually a hopper.
This fixes a bug that led to abandoned ArmorStands (what I was referring
to when I said "target") that lingered after the hopper was destroyed.
The issue was that tick (called asynchronously) was creating a new
ArmorStand (via InfusedHopper.getArmorStand()) after the block handler's
onBreak call removed the one that was created in onPlace.

The code that recreates the ArmorStand remains in tick() just in case,
somehow, it ends up getting removed by some other force. However, this
code should never be reached if the hopper itself is removed.

This fixes TheBusyBiscuit/Slimefun4#189,
fixes TheBusyBiscuit/Slimefun4#190, and
fixes TheBusyBiscuit/Slimefun4#428.