Do not overwrite block for a location that already has a valid block
This shouldn't happen, but if it does happen it'll definitely break the block.
We'll notify the server administrator on load and only keep the first block we
see.

Note that this will only happen if there's something seriously wrong with the
block as generally there'll not be a situation where information is stored in
two different files.