Do not allow more than 54 items in reactor/generator setup guide slots
If we don't, opening the schematics for a Bio Reactor results in

"IllegalArgumentException: Size for custom inventory must be a multiple
 of 9 between 9 and 54 slots"

And the player is simply handed the Bio Reactor, which they can place
and use. Once they do that, the Slimefun guide can't be used until they
reconnect.

For generators and reactors we show the fuels underneath the schematics
or build instructions. When there are more fuels than available space to
show them, we'll overflow, causing the custom inventory to not work and
Slimefun throwing exceptions.