Damage explosive pickaxe on use (configurable, off by default)
The Explosive Pickaxe currently doesn't take any damage and therefore
doesn't lose durability, which isn't a desirable behavior for some of
us.

This commit adds a 'damage-on-use' item setting for EXPLOSIVE_PICKAXE,
which is disabled by default. When enabled, the Explosive Pickaxe will
get damaged for each block it breaks, in the same way a normal Diamond
Pickaxe would get damage if it were to break those blocks. This seems
to be similar to how Slimefun damages other tools.

Players can still enable Unbreaking (e.g. durability) on the pickaxe.