Drop Slimefun items when broken by another Slimefun tool
When a Slimefun item that doesn't have a BlockHandler is broken by a
tool that does have a BlockHandler and that tool's BlockHandler returns
true, Slimefun would stop handling the "onBlockBreak" event and would
not drop the Slimefun item that was broken but some vanilla block
instead.

This commit attempts to fix the way those events are handled, ensuring
that if a Slimefun item is broken one is dropped for the player.