REACTOR/PORT UPDATE
Update Reactor and Reactor Access Port to be a bit easier to use, now
when a reactor detects an attached access port the GUI will be updated
to show the two are linked.  Once linked the reactor access port will
show the status of the actual reactor rather than it's own inventory,
allowing users to remotely control the reactor's priority and see the
remaining fuel/coolant.   Items can still be added to the access port
manually or sent in from the cargo manager.