Fix deadlock when shutting down
The `processSyncTasks` method had a bug in it which caused an infinite loop. This was running on the main thread (since sync) and thus never terminated.

On shutdown we call `tick()`, this will submit async tasks and process the final sync tasks. In `processSyncTasks` we have a while loop. `while (!plugin.isEnabled() || deadline > System.nanoTime()) {` the plugin is shutting down therefore `isEnabled` will return false meaning this while loop will always trigger. Next we polled the tasks (`syncTasks.poll(MAX_POLL_TIME, TimeUnit.NANOSECONDS);`). When there are none left this will return null. We have a null check and inside that block we haev `if (plugin.isEnabled()) { break; }`. We want to break out of this loop however as mentioned already `isEnabled` will return false and thus causing this infinite loop bug.