Fixes #3225 and chunk data not being saved on shutdown
We had a `chunksLoaded` boolean, not sure why this was here but removing it did not cause any unexpected changes.

We have a world name check so this in theory should not be required. If there was a reason which is still valid, we will find out soon and should do a proper fix rather than this.

I also fixed the chunk data not saving on normal shutdown