Fix test assertions and test new change
The tests were previously always clearing BlockStorage after break due to location re-use. This is generally a bad practice anyway and broke this since we now explictly want to ensure that we cannot place if a deletion is queued.

Added a new test which will assert that after placing and breaking, you cannot now place another before it is fully deleted.

Co-authored-by: J3fftw <44972470+J3fftw1@users.noreply.github.com>