:sparkles: feat(api): Introduce SlimefunRegistryFinalizedEvent
This Event is provided so that plugins that have integrations with each other aren't
stuck in a state in which some SlimefunItems may not be available to one plugin or the other.

I will use this event to provide a cleaner integration between Dynatech and other plugins.
Currently InfinityXpansion, ExoticGarden, and Gastronomicon.