Move PlayerProfile saving off the main thread (#4119)
* Move PlayerProfile off main thread, add debugs and improve tab completion for debug

Moved the PlayerProfile saving off the main thread, we generally load this off-thread but now we also save off-thread. I thought we were already doing this but apparently not, especially with our current YAML stuff this should definitely be done

Also done a small change to ensure that we don't remove the PlayerProfile from memory if the player is still online. I don't think we ever had a reported issue from this but it's kinda weird behaviour

Finally, added some debug logs to the saving logic, this can be enabled with `sf debug slimefun_player_profile_data`. Also added auto-complete to /sf debug because it's nice, this only works for Slimefun test cases rather than addons but that's fine. Mostly internal anyway

* Update src/main/java/io/github/thebusybiscuit/slimefun4/api/player/PlayerProfile.java

---------

Co-authored-by: Alessio Colombo <37039432+Sfiguz7@users.noreply.github.com>